requirements.txt
yfinance
pandas
ta
litellm
feedparser
python-dotenv
alpaca-py
pytest
pyyaml